2014-07-03 12:08 UTC+0300 Alexander Kresin <alex@belacy.ru>   * source/gtk/hbrowse.prg   * source/winapi/hbrowse.prg     * Added checking of existing a HBitmap object in HColumn():aBitmaps   * source/common/xml/xmlparse.c   * source/common/editor/hcedit_w.c   * source/common/editor/hcedit_l.c     ! Warnings fixed, reported by Alain Aupeix and Maurizio la Cecilia   * source/common/editor/hcedit.prg     * Added support for Ctrl+C, Ctrl+X, Ctrl+V and Ctrl+Z   * utils/designer/designer.prg   * utils/designer/hformgen.prg     * You can now insert a following item to designer.iml:       <rep_zoom default="1.2"/>       It cause zooming of a report window ( +20% in above case ), just choose       a value, most comfortable for you.   * utils/designer/editor.prg     * Added "Edit" menu to the editor dialog box
git-svn-id: https://svn.code.sf.net/p/hwgui/code/trunk/hwgui@2274 66345640-f411-406e-acba-706ed0d45c9c

+++ source/common/xml/xmlparse.c
@@ -236,7 +236,7 @@ HB_FUNC( HBXML_PRELOAD )
    unsigned char *ucSource = (unsigned char *)hb_parc(1);
    unsigned char *ptr = ucSource;
    unsigned long ulNew = 0;
-   unsigned long ulLen = (unsigned long)((HB_ISNUM(2))? hb_parnl(2) : hb_parclen(1));
+   unsigned long ulLen = ((HB_ISNUM(2))? (unsigned long)hb_parnl(2) : (unsigned long)hb_parclen(1));
    unsigned char *ptrnew = (unsigned char *) malloc( ulLen+1 );
    int i, nlen;
    int iChar;
@@ -715,4 +715,3 @@ HB_FUNC( HBXML_GETDOC )
 
    hb_retni( nParseError );
 }
-
